Poincare Map Video
The Poincaré map is generated by numerically solving the differential equation $\frac{dx}{dt} = x - x^3 - b\sin(2\pi t)$
using the `solve-ivp' function from the SciPy library for different values of $b$. We solve the initial value problem for $0 \le t \le 1$ with a range of initial values for $x(0) \in [-2,2]$ and plot the solution points. Then we plot the $y=x$ line and we know from our text that the solution points which will intersect with this line- will give us the periodic solutions and then we generate a dynamic plot-gif which plots these solutions points to visualize the system's behavior as $b$ varied.